FEAT: add auth guard to some routes

+++ src/handlers/people/teachers.rs
@@ -1,7 +1,10 @@
-use crate::types::shared::{ResponseData, State};
+use crate::{
+    guards::auth_guard,
+    types::shared::{ResponseData, State},
+};
 use actix_web::{
     web::{Json, Path, Query},
-    HttpResponse as Response,
+    HttpRequest as Request, HttpResponse as Response,
 };
 use service::{
     models::{Teacher, TeacherQuery},
@@ -10,7 +13,19 @@ use service::{
     uuid::Uuid,
 };
 //
-pub async fn create(body: Json<Teacher>, state: State) -> Response {
+pub async fn create(req: Request, body: Json<Teacher>, state: State) -> Response {
+    // get headers
+    let headers = req.headers();
+    // check token for auth
+    let authorized = auth_guard(headers, state.config.jwt_secret.clone());
+    // unauth
+    if let Err(message) = authorized {
+        return Response::Unauthorized().json(ResponseData::<String> {
+            error: Some(message),
+            message: None,
+            data: None,
+        });
+    }
     let res = MutationService::create_teacher(&state.db_conn, body.into_inner()).await;
     match res {
         Ok(id) => Response::Ok().json(ResponseData {
@@ -26,7 +41,19 @@ pub async fn create(body: Json<Teacher>, state: State) -> Response {
     }
 }
 
-pub async fn add_subject(params: Path<(Uuid, Uuid)>, state: State) -> Response {
+pub async fn add_subject(req: Request, params: Path<(Uuid, Uuid)>, state: State) -> Response {
+    // get headers
+    let headers = req.headers();
+    // check token for auth
+    let authorized = auth_guard(headers, state.config.jwt_secret.clone());
+    // unauth
+    if let Err(message) = authorized {
+        return Response::Unauthorized().json(ResponseData::<String> {
+            error: Some(message),
+            message: None,
+            data: None,
+        });
+    }
     let res = MutationService::create_teacher_subject(&state.db_conn, params.into_inner()).await;
     match res {
         Ok(id) => Response::Ok().json(ResponseData {
@@ -42,7 +69,19 @@ pub async fn add_subject(params: Path<(Uuid, Uuid)>, state: State) -> Response {
     }
 }
 
-pub async fn delete_subject(id: Path<Uuid>, state: State) -> Response {
+pub async fn delete_subject(req: Request, id: Path<Uuid>, state: State) -> Response {
+    // get headers
+    let headers = req.headers();
+    // check token for auth
+    let authorized = auth_guard(headers, state.config.jwt_secret.clone());
+    // unauth
+    if let Err(message) = authorized {
+        return Response::Unauthorized().json(ResponseData::<String> {
+            error: Some(message),
+            message: None,
+            data: None,
+        });
+    }
     let res = MutationService::delete_teacher_subject(&state.db_conn, id.into_inner()).await;
     match res {
         Ok(delete_count) => Response::Ok().json(ResponseData {
@@ -58,7 +97,19 @@ pub async fn delete_subject(id: Path<Uuid>, state: State) -> Response {
     }
 }
 
-pub async fn delete(id: Path<Uuid>, state: State) -> Response {
+pub async fn delete(req: Request, id: Path<Uuid>, state: State) -> Response {
+    // get headers
+    let headers = req.headers();
+    // check token for auth
+    let authorized = auth_guard(headers, state.config.jwt_secret.clone());
+    // unauth
+    if let Err(message) = authorized {
+        return Response::Unauthorized().json(ResponseData::<String> {
+            error: Some(message),
+            message: None,
+            data: None,
+        });
+    }
     let res = MutationService::delete_teacher(&state.db_conn, id.into_inner()).await;
     match res {
         Ok(delete_count) => Response::Ok().json(ResponseData {
@@ -74,7 +125,19 @@ pub async fn delete(id: Path<Uuid>, state: State) -> Response {
     }
 }
 
-pub async fn list(query: Query<TeacherQuery>, state: State) -> Response {
+pub async fn list(req: Request, query: Query<TeacherQuery>, state: State) -> Response {
+    // get headers
+    let headers = req.headers();
+    // check token for auth
+    let authorized = auth_guard(headers, state.config.jwt_secret.clone());
+    // unauth
+    if let Err(message) = authorized {
+        return Response::Unauthorized().json(ResponseData::<String> {
+            error: Some(message),
+            message: None,
+            data: None,
+        });
+    }
     let res = QueryService::list_teachers(&state.db_conn, query.into_inner()).await;
     match res {
         Ok(teachers) => Response::Ok().json(ResponseData {
@@ -90,7 +153,19 @@ pub async fn list(query: Query<TeacherQuery>, state: State) -> Response {
     }
 }
 
-pub async fn update(id: Path<Uuid>, body: Json<Teacher>, state: State) -> Response {
+pub async fn update(req: Request, id: Path<Uuid>, body: Json<Teacher>, state: State) -> Response {
+    // get headers
+    let headers = req.headers();
+    // check token for auth
+    let authorized = auth_guard(headers, state.config.jwt_secret.clone());
+    // unauth
+    if let Err(message) = authorized {
+        return Response::Unauthorized().json(ResponseData::<String> {
+            error: Some(message),
+            message: None,
+            data: None,
+        });
+    }
     let res =
         MutationService::update_teacher(&state.db_conn, id.into_inner(), body.into_inner()).await;
     match res {
